pyproject.toml
[project]
name = "django-podcast-analyzer"
version = "0.0.1"
description = "A simple Django app that tracks podcast feeds and provides some useful metrics about them."
authors = [{ name = "Daniel Andrlik", email = "daniel@andrlik.org" }]
license = "BSD-3-Clause"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-q2>=1.7.2",
    "django-rich>=1.11.0",
    "django-tagulous>=2.1.0",
    "django-typer>=2.2.2",
    "django>=5.0.0",
    "httpx[brotli,http2]>=0.27.2",
    "pillow>=10.4.0",
    "podcastparser>=0.6.10",
    "python-magic>=0.4.27",
    "urlman>=2.0.2",
]
classifiers = [
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "License :: OSI Approved :: BSD License",
    "Framework :: Django",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Repository = "https://github.com/andrlik/django-podcast-analyzer"
Homepage = "https://github.com/andrlik/django-podcast-analyzer"
Documentation = "https://andrlik.github.io/django-podcast-analyzer/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/podcast_analyzer"]

[tool.hatch.build.targets.sdist]
include = ["src/podcast_analyzer", "tests"]

[tool.uv]
dev-dependencies = [
    "bandit>=1.7.10",
    "blessed>=1.20.0",
    "bump-my-version>=0.27.0",
    "django-browser-reload>=1.15.0",
    "django-coverage-plugin>=3.1.0",
    "django-debug-toolbar>=4.4.6",
    "django-extensions>=3.2.3",
    "django-types>=0.19.1",
    "django-watchfiles>=0.2.0",
    "factory-boy>=3.3.1",
    "mike>=2.1.3",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-include-markdown-plugin>=6.2.2",
    "mkdocs-literate-nav>=0.6.1",
    "mkdocs-material>=9.5.40",
    "mkdocstrings[python]>=0.26.1",
    "pyright>=1.1.384",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-django>=4.9.0",
    "pytest-httpx>=0.32.0",
    "pytest-sugar>=1.0.0",
    "ruff>=0.6.9",
    "werkzeug[watchdog]>=3.0.4",
]

[tool.coverage.run]
source = ["src/podcast_analyzer"]
omit = ["*migrations*", "*tests*", "*staticfiles*"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.black]
# https://github.com/psf/black
target-version = ["py312"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | migrations
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.ruff]
target-version = "py312"
line-length = 88
extend-exclude = ["migrations", ".tox"]

[tool.ruff.lint]
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DJ",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105",
  "S106",
  "S107",
  # Ignore complexity
  "C901",
  "PLR0911",
  "PLR0912",
  "PLR0913",
  "PLR0915",
  "DJ001",
  "RUF012",
  "ISC001",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101", "FBT001", "FBT002", "ARG001", "ARG002", "E501", "PLR2004", "T201"]
"tests/urls.py" = ["RUF005"]
"src/podcast_analyzers/receivers.py" = ["ARG001"]
"src/podcast_analyzer/rules.py" = ["ARG001"]
"src/podcast_analyzer/views.py" = ["A001"]
"src/podcast_analyzer/management/commands/*.py" = ["ARG002"]
"conftest.py" = ["ARG001"]

[tool.ruff.lint.isort]
known-first-party = ["podcast_analyzer"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pyright]
pythonVersion = "3.12"
exclude = ["**/migrations/",
    "**/__pycache__",
    ".tox",
    ".venv",
    "tests",
    "conftest.py",
]
reportIncompatibleVariableOverride = false
reportIncompatibleMethodOverride = false

[tool.django-stubs]
django_settings_module = "tests.settings"

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
DJANGO_SETTINGS_MODULE = "tests.settings"
testpaths = ["tests"]
pythonpath = [".", "src"]
django_find_project = false
asyncio_default_fixture_loop_scope = "function"
norecursedirs = ["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__", ".venv", ".tox"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::PendingDeprecationWarning"
]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules", "--cov=src/podcast_analyzer",
  "--cov-report=term-missing", "--ds=tests.settings", "--reuse-db",
]

[tool.check-wheel-contents]
ignore = ["W004"]

[tool.curlylint.rules]
image_alt = true

[tool.bumpversion]
current_version = "0.0.1"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "src/podcast_analyzer/__init__.py"
search = "__version__ = \"{current_version}\""
replace =  "__version__ = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = "version = \"{current_version}\""
replace = "version = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = "Unreleased"

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = "{current_version}...HEAD"
replace = "{current_version}...{new_version}"
